Convert User Scripts to Python
This commits converts all of the previous bash scripts to python.
This was done to make the handling of usernames and passwords more
secure with the use of python's subprocess library. This will also
allow the passwords for users to be more secure since the users
input does not have to be sanitized now.